+++ xrcf/tests/convert/mod.rs
@@ -2,4 +2,5 @@ mod cf_to_llvm;
 mod experimental_to_mlir;
 mod func_to_llvm;
 mod mlir_to_llvmir;
+mod mlir_to_wat;
 mod scf_to_cf;
